+++ src/amo.ts
@@ -1,8 +1,5 @@
-import FormData from "form-data";
-import fetch from "node-fetch";
-import { v4 as uuidv4 } from "uuid";
-import type { ReadStream } from "fs";
 import jwt from "jsonwebtoken";
+import { v4 as uuidv4 } from "uuid";
 
 type VersionRange = { min?: string; max?: string };
 type Compatibility = Record<string, VersionRange> | Array<string>;
@@ -120,12 +117,12 @@ export class AMOClient {
   }
 
   async uploadAddon(
-    xpi: ReadStream,
-    channel: Channel
+    xpi: Blob,
+    channel: Channel,
   ): Promise<AMOApiUploadDetailResponse> {
-    const path = `/api/v5/addons/upload/`;
+    const path = "/api/v5/addons/upload/";
     const form = new FormData();
-    form.append("upload", xpi);
+    form.append("upload", xpi, "addon.zip");
     form.append("channel", channel);
 
     return this.proceed<AMOApiUploadDetailResponse>(path, "POST", form);
@@ -140,12 +137,12 @@ export class AMOClient {
   async uploadSource(
     addon: number | string,
     version: string,
-    source: ReadStream,
-    license?: License
+    source: Blob,
+    license?: License,
   ): Promise<VersionDetailResponse> {
     const path = `/api/v5/addons/addon/${addon}/versions/${version}/`;
     const form = new FormData();
-    form.append("source", source);
+    form.append("source", source, "source.zip");
     form.append("license", license);
 
     return this.proceed<VersionDetailResponse>(path, "PATCH", form);
@@ -177,7 +174,7 @@ export class AMOClient {
 
   async getVersionOrUndefined(
     addon: number | string,
-    version: number | string
+    version: number | string,
   ): Promise<VersionDetailResponse | undefined> {
     const path = `/api/v5/addons/addon/${addon}/versions/${version}/`;
 
@@ -187,12 +184,12 @@ export class AMOClient {
   private async proceed<T>(
     path: string,
     method: string,
-    params?: unknown
+    params?: unknown,
   ): Promise<T> {
     const token = this._getJwtToken();
     const url = `${this.origin}${path}`;
-    const headers: Record<string, string> = { Authorization: "JWT " + token };
-    let body;
+    const headers: Record<string, string> = { Authorization: `JWT ${token}` };
+    let body: string | FormData | undefined;
 
     if (params instanceof FormData) {
       body = params;
@@ -208,7 +205,7 @@ export class AMOClient {
       throw new Error(
         `Failed to ${method} ${url}: ${resp.status} ${
           resp.statusText
-        } ${await resp.text()}`
+        } ${await resp.text()}`,
       );
     }
 
@@ -218,12 +215,12 @@ export class AMOClient {
   private async proceedOrUndefined<T>(
     path: string,
     method: string,
-    params?: unknown
+    params?: unknown,
   ): Promise<T | undefined> {
     const token = this._getJwtToken();
     const url = `${this.origin}${path}`;
-    const headers: Record<string, string> = { Authorization: "JWT " + token };
-    let body;
+    const headers: Record<string, string> = { Authorization: `JWT ${token}` };
+    let body: string | FormData | undefined;
 
     if (params instanceof FormData) {
       body = params;
@@ -242,7 +239,7 @@ export class AMOClient {
       throw new Error(
         `Failed to ${method} ${url}: ${resp.status} ${
           resp.statusText
-        } ${await resp.text()}`
+        } ${await resp.text()}`,
       );
     }